Functions related to bridging and frames.

See also countCompetitors, trackPrevContext.

Usage

lastBridgeUnit(
  frameMatrix,
  unitSeq = NULL,
  chain = NULL,
  tokenOrderFirst = NULL,
  tokenOrderLast = NULL,
  inclRelations = NULL
)

unitsToLastBridge(
  frameMatrix,
  unitSeq = NULL,
  chain = NULL,
  tokenOrderLast = NULL,
  tokenOrderFirst = NULL,
  inclRelations = NULL
)

lastBridgeToken(
  frameMatrix,
  firstOrLast = "last",
  tokenOrderFirst = NULL,
  tokenOrderLast = NULL,
  chain = NULL,
  inclRelations = NULL
)

tokensToLastBridge(
  frameMatrix,
  firstOrLast = "last",
  tokenOrderFirst = NULL,
  tokenOrderLast = NULL,
  chain = NULL,
  zeroProtocol = "literal",
  zeroCond = NULL,
  unitSeq = NULL,
  unitDF = NULL,
  inclRelations = NULL
)

Arguments

unitSeq

The vector of units where the mentions appeared.

chain

The chain that each mention belongs to.

zeroProtocol

If 'literal', I will take the seq values of the zeroes at face value. (If you set zeros as non-words and use docWordSeqFirst or discourseWordSeLast as your tokenOrder, this will lead to meaningless values.) If 'unitFinal', I will treat zeroes as if they were the final word of the unit. If 'unitFirst', I will treat zeroes as if they were the first word of the unit.

zeroCond

A condition for determining whether a token is zero. For most people, this should be (word column) == "<0>".

tokenOrder

The vector of sequence values values where the mentions appeared. Common choices are docTokenSeqFirst, docTokenSeqLast, wordTokenSeqFirst and wordTokenseqLast (the last two are available after running addIsWordField on a rezrObj. By default it's docTokenSeqLast.