Read Pangloss file

Read Pangloss file

Usage

rez_pangloss(url, filename)

Arguments

url

A URL for the Pangloss source.

filename

a filename for the export.

Value

rezImport

Examples

rez_pangloss("https://cocoon.huma-num.fr/data/mazaudon/masters/crdo-TAJ_ORIGTAM.xml", "tamang1.txt")
#>     word_id text_id sentence_id       word_form word_gloss sent_audio_start
#> 1         1       1   ORIGTAMs1           ⁴unma       <NA>           0.0000
#> 2         2       1   ORIGTAMs1          ⁴taŋpo       <NA>           0.0000
#> 3         3       1   ORIGTAMs1         ⁴tui-ri       <NA>           0.0000
#> 4         4       1   ORIGTAMs1         ¹tor-se       <NA>           0.0000
#> 5         5       1   ORIGTAMs1         ³poi-ri       <NA>           0.0000
#> 6         6       1   ORIGTAMs2           ⁴unma       <NA>           5.7573
#> 7         7       1   ORIGTAMs2          ⁴taŋpo       <NA>           5.7573
#> 8         8       1   ORIGTAMs2         ⁴tui-ri       <NA>           5.7573
#> 9         9       1   ORIGTAMs2         ¹tor-se       <NA>           5.7573
#> 10       10       1   ORIGTAMs2      ³poi-ri-mi       <NA>           5.7573
#> 11       11       1   ORIGTAMs3          chetri       <NA>          11.0001
#> 12       12       1   ORIGTAMs3            ⁴kik       <NA>          11.0001
#> 13       13       1   ORIGTAMs3         ⁴pa:man       <NA>          11.0001
#> 14       14       1   ORIGTAMs3            ⁴kik       <NA>          11.0001
#> 15       15       1   ORIGTAMs3          ¹tamaŋ       <NA>          11.0001
#> 16       16       1   ORIGTAMs3            ⁴kik       <NA>          11.0001
#> 17       17       1   ORIGTAMs3     ²som-cim-ro       <NA>          11.0001
#> 18       18       1   ORIGTAMs4     ²som-naŋ-ri       <NA>          18.3013
#> 19       19       1   ORIGTAMs5         ¹ama-mi       <NA>          22.1610
#> 20       20       1   ORIGTAMs5          ³palaŋ       <NA>          22.1610
#> 21       21       1   ORIGTAMs5          ¹apa-m       <NA>          22.1610
#> 22       22       1   ORIGTAMs5 ¹hanuman-cim-ro       <NA>          22.1610
#> 23       23       1   ORIGTAMs6         ¹am'-se       <NA>          32.0675
#> 24       24       1   ORIGTAMs6            ¹tik       <NA>          32.0675
#> 25       25       1   ORIGTAMs6      ³pi-cim-ro       <NA>          32.0675
#> 26       26       1   ORIGTAMs7        ¹thetmai       <NA>          36.8599
#> 27       27       1   ORIGTAMs7            ²som       <NA>          36.8599
#> 28       28       1   ORIGTAMs7          ³ca-ta       <NA>          36.8599
#> 29       29       1   ORIGTAMs8          ⁴camam       <NA>          41.4915
#> 30       30       1   ORIGTAMs8             ¹ŋa       <NA>          41.4915
#> 31       31       1   ORIGTAMs8          ¹si-ma       <NA>          41.4915
#> 32       32       1   ORIGTAMs8             ¹ŋa       <NA>          41.4915
#> 33       33       1   ORIGTAMs8          ¹si-ma       <NA>          41.4915
#> 34       34       1   ORIGTAMs8         bel'-ri       <NA>          41.4915
#> 35       35       1   ORIGTAMs8             ¹ŋa       <NA>          41.4915
#> 36       36       1   ORIGTAMs8          ¹si-ma       <NA>          41.4915
#> 37       37       1   ORIGTAMs9          ¹ŋa-la       <NA>          51.5267
#> 38       38       1   ORIGTAMs9            ¹sja       <NA>          51.5267
#> 39       39       1   ORIGTAMs9       ²ai-ni-se       <NA>          51.5267
#> 40       40       1   ORIGTAMs9    ³a-²khraŋ-na       <NA>          51.5267
#> 41       41       1   ORIGTAMs9     ³a-⁴pjaŋ-na       <NA>          51.5267
#> 42       42       1  ORIGTAMs10          ¹ŋa-la       <NA>          57.9595
#> 43       43       1  ORIGTAMs10            ¹sja       <NA>          57.9595
#> 44       44       1  ORIGTAMs10          ²ai-ni       <NA>          57.9595
#> 45       45       1  ORIGTAMs10   ²kola-kat'-se       <NA>          57.9595
#> 46       46       1  ORIGTAMs10           ¹ca-o       <NA>          57.9595
#> 47       47       1  ORIGTAMs10      ³pi-cim-ro       <NA>          57.9595
#> 48       48       1  ORIGTAMs11          ³pi-ma       <NA>          62.3338
#> 49       49       1  ORIGTAMs11            ³oca       <NA>          62.3338
#> 50       50       1  ORIGTAMs11            ¹ama       <NA>          62.3338
#> 51       51       1  ORIGTAMs11          ¹si-si       <NA>          62.3338
#> 52       52       1  ORIGTAMs11          ¹ni-ci       <NA>          62.3338
#> 53       53       1  ORIGTAMs12             ³ca       <NA>          68.6379
#> 54       54       1  ORIGTAMs12         ²som-se       <NA>          68.6379
#> 55       55       1  ORIGTAMs12          ¹am-ta       <NA>          68.6379
#> 56       56       1  ORIGTAMs12       ¹phral-si       <NA>          68.6379
#> 57       57       1  ORIGTAMs12        ²tha:-si       <NA>          68.6379
#> 58       58       1  ORIGTAMs12          ¹am-ta       <NA>          68.6379
#> 59       59       1  ORIGTAMs12       ¹phral-si       <NA>          68.6379
#> 60       60       1  ORIGTAMs12        ²tha:-si       <NA>          68.6379
#> 61       61       1  ORIGTAMs13             ³ca       <NA>          76.7915
#> 62       62       1  ORIGTAMs13          ¹thepa       <NA>          76.7915
#> 63       63       1  ORIGTAMs13           (aaa)       <NA>          76.7915
#> 64       64       1  ORIGTAMs14          ¹thepa       <NA>          79.3255
#> 65       65       1  ORIGTAMs14       ²paraŋ-ca       <NA>          79.3255
#> 66       66       1  ORIGTAMs14        ²tha:-ci       <NA>          79.3255
#> 67       67       1  ORIGTAMs15             ³ca       <NA>          83.0705
#> 68       68       1  ORIGTAMs15      ¹cjaŋpa-ca       <NA>          83.0705
#> 69       69       1  ORIGTAMs15           ³kruŋ       <NA>          83.0705
#> 70       70       1  ORIGTAMs15     ²khru-pa-r'       <NA>          83.0705
#> 71       71       1  ORIGTAMs15     ²pit-cim-ro       <NA>          83.0705
#> 72       72       1  ORIGTAMs16             ³ca       <NA>          87.5178
#> 73       73       1  ORIGTAMs16      ¹cjaŋpa-ca       <NA>          87.5178
#> 74       74       1  ORIGTAMs16           ³kruŋ       <NA>          87.5178
#> 75       75       1  ORIGTAMs16     ²khru-pa-ri       <NA>          87.5178
#> 76       76       1  ORIGTAMs16         ²pit-ma       <NA>          87.5178
#> 77       77       1  ORIGTAMs17           ³kruŋ       <NA>          93.0925
#> 78       78       1  ORIGTAMs17        ²khru-si       <NA>          93.0925
#> 79       79       1  ORIGTAMs17          ³pa-ma       <NA>          93.0925
#> 80       80       1  ORIGTAMs18         ³cjocjo       <NA>          95.2007
#> 81       81       1  ORIGTAMs18         ⁴ni:-se       <NA>          95.2007
#> 82       82       1  ORIGTAMs18            ¹tik       <NA>          95.2007
#> 83       83       1  ORIGTAMs18      ¹la-cim-ro       <NA>          95.2007
#> 84       84       1  ORIGTAMs18         ³pi-sam       <NA>          95.2007
#> 85       85       1  ORIGTAMs19            ¹sja       <NA>          98.9281
#> 86       86       1  ORIGTAMs19       ²khraŋ-si       <NA>          98.9281
#> 87       87       1  ORIGTAMs19           (bis)       <NA>          98.9281
#> 88       88       1  ORIGTAMs19            ³ol'       <NA>          98.9281
#> 89       89       1  ORIGTAMs19          ¹la-si       <NA>          98.9281
#> 90       90       1  ORIGTAMs19             bag       <NA>          98.9281
#> 91       91       1  ORIGTAMs19            ²som       <NA>          98.9281
#> 92       92       1  ORIGTAMs19          ¹la-si       <NA>          98.9281
#> 93       93       1  ORIGTAMs20         ³cjocjo       <NA>         106.4759
#> 94       94       1  ORIGTAMs20         ⁴ni:-la       <NA>         106.4759
#> 95       95       1  ORIGTAMs20       ¹sja-cati       <NA>         106.4759
#> 96       96       1  ORIGTAMs20            ²tok       <NA>         106.4759
#> 97       97       1  ORIGTAMs20            ²sum       <NA>         106.4759
#> 98       98       1  ORIGTAMs20    ¹than-cim-ro       <NA>         106.4759
#> 99       99       1  ORIGTAMs21           ³kruŋ       <NA>         108.6520
#> 100     100       1  ORIGTAMs21        ²khru-si       <NA>         108.6520
#> 101     101       1  ORIGTAMs21         ¹kha-pa       <NA>         108.6520
#> 102     102       1  ORIGTAMs21   ¹cjaŋpa-ca-la       <NA>         108.6520
#> 103     103       1  ORIGTAMs21            ¹sja       <NA>         108.6520
#> 104     104       1  ORIGTAMs22            ¹ale       <NA>         111.2559
#> 105     105       1  ORIGTAMs22            ¹ale       <NA>         111.2559
#> 106     106       1  ORIGTAMs22       ¹in-che-m       <NA>         111.2559
#> 107     107       1  ORIGTAMs22          ³tolon       <NA>         111.2559
#> 108     108       1  ORIGTAMs22          ¹ca-ci       <NA>         111.2559
#> 109     109       1  ORIGTAMs23          ²ai-la       <NA>         114.0249
#> 110     110       1  ORIGTAMs23             bag       <NA>         114.0249
#> 111     111       1  ORIGTAMs23          ²cu-ri       <NA>         114.0249
#>     sent_audio_end
#> 1           5.7573
#> 2           5.7573
#> 3           5.7573
#> 4           5.7573
#> 5           5.7573
#> 6          11.0001
#> 7          11.0001
#> 8          11.0001
#> 9          11.0001
#> 10         11.0001
#> 11         18.3013
#> 12         18.3013
#> 13         18.3013
#> 14         18.3013
#> 15         18.3013
#> 16         18.3013
#> 17         18.3013
#> 18         22.1610
#> 19         32.0675
#> 20         32.0675
#> 21         32.0675
#> 22         32.0675
#> 23         36.8599
#> 24         36.8599
#> 25         36.8599
#> 26         41.4915
#> 27         41.4915
#> 28         41.4915
#> 29         51.5267
#> 30         51.5267
#> 31         51.5267
#> 32         51.5267
#> 33         51.5267
#> 34         51.5267
#> 35         51.5267
#> 36         51.5267
#> 37         57.9595
#> 38         57.9595
#> 39         57.9595
#> 40         57.9595
#> 41         57.9595
#> 42         62.3338
#> 43         62.3338
#> 44         62.3338
#> 45         62.3338
#> 46         62.3338
#> 47         62.3338
#> 48         68.6379
#> 49         68.6379
#> 50         68.6379
#> 51         68.6379
#> 52         68.6379
#> 53         76.7915
#> 54         76.7915
#> 55         76.7915
#> 56         76.7915
#> 57         76.7915
#> 58         76.7915
#> 59         76.7915
#> 60         76.7915
#> 61         79.3255
#> 62         79.3255
#> 63         79.3255
#> 64         82.6838
#> 65         82.6838
#> 66         82.6838
#> 67         86.1134
#> 68         86.1134
#> 69         86.1134
#> 70         86.1134
#> 71         86.1134
#> 72         90.6475
#> 73         90.6475
#> 74         90.6475
#> 75         90.6475
#> 76         90.6475
#> 77         95.2007
#> 78         95.2007
#> 79         95.2007
#> 80         98.1566
#> 81         98.1566
#> 82         98.1566
#> 83         98.1566
#> 84         98.1566
#> 85        105.6874
#> 86        105.6874
#> 87        105.6874
#> 88        105.6874
#> 89        105.6874
#> 90        105.6874
#> 91        105.6874
#> 92        105.6874
#> 93        108.6520
#> 94        108.6520
#> 95        108.6520
#> 96        108.6520
#> 97        108.6520
#> 98        108.6520
#> 99        110.6233
#> 100       110.6233
#> 101       110.6233
#> 102       110.6233
#> 103       110.6233
#> 104       113.2639
#> 105       113.2639
#> 106       113.2639
#> 107       113.2639
#> 108       113.2639
#> 109       115.5469
#> 110       115.5469
#> 111       115.5469
#>                                                                sent_form
#> 1                                  ⁴unma ⁴taŋpo ⁴tui-ri, ¹tor-se ³poi-ri
#> 2                                  ⁴unma ⁴taŋpo ⁴tui-ri, ¹tor-se ³poi-ri
#> 3                                  ⁴unma ⁴taŋpo ⁴tui-ri, ¹tor-se ³poi-ri
#> 4                                  ⁴unma ⁴taŋpo ⁴tui-ri, ¹tor-se ³poi-ri
#> 5                                  ⁴unma ⁴taŋpo ⁴tui-ri, ¹tor-se ³poi-ri
#> 6                               ⁴unma ⁴taŋpo ⁴tui-ri, ¹tor-se ³poi-ri-mi
#> 7                               ⁴unma ⁴taŋpo ⁴tui-ri, ¹tor-se ³poi-ri-mi
#> 8                               ⁴unma ⁴taŋpo ⁴tui-ri, ¹tor-se ³poi-ri-mi
#> 9                               ⁴unma ⁴taŋpo ⁴tui-ri, ¹tor-se ³poi-ri-mi
#> 10                              ⁴unma ⁴taŋpo ⁴tui-ri, ¹tor-se ³poi-ri-mi
#> 11                   chetri ⁴kik, ⁴pa:man ⁴kik, ¹tamaŋ ⁴kik, ²som-cim-ro
#> 12                   chetri ⁴kik, ⁴pa:man ⁴kik, ¹tamaŋ ⁴kik, ²som-cim-ro
#> 13                   chetri ⁴kik, ⁴pa:man ⁴kik, ¹tamaŋ ⁴kik, ²som-cim-ro
#> 14                   chetri ⁴kik, ⁴pa:man ⁴kik, ¹tamaŋ ⁴kik, ²som-cim-ro
#> 15                   chetri ⁴kik, ⁴pa:man ⁴kik, ¹tamaŋ ⁴kik, ²som-cim-ro
#> 16                   chetri ⁴kik, ⁴pa:man ⁴kik, ¹tamaŋ ⁴kik, ²som-cim-ro
#> 17                   chetri ⁴kik, ⁴pa:man ⁴kik, ¹tamaŋ ⁴kik, ²som-cim-ro
#> 18                                                           ²som-naŋ-ri
#> 19                                ¹ama-mi ³palaŋ, ¹apa-m ¹hanuman-cim-ro
#> 20                                ¹ama-mi ³palaŋ, ¹apa-m ¹hanuman-cim-ro
#> 21                                ¹ama-mi ³palaŋ, ¹apa-m ¹hanuman-cim-ro
#> 22                                ¹ama-mi ³palaŋ, ¹apa-m ¹hanuman-cim-ro
#> 23                                               ¹am'-se ¹tik ³pi-cim-ro
#> 24                                               ¹am'-se ¹tik ³pi-cim-ro
#> 25                                               ¹am'-se ¹tik ³pi-cim-ro
#> 26                                                ¹thetmai ²som ³ca-ta, 
#> 27                                                ¹thetmai ²som ³ca-ta, 
#> 28                                                ¹thetmai ²som ³ca-ta, 
#> 29                     ⁴camam ¹ŋa ¹si-ma, ¹ŋa ¹si-ma bel'-ri, ¹ŋa ¹si-ma
#> 30                     ⁴camam ¹ŋa ¹si-ma, ¹ŋa ¹si-ma bel'-ri, ¹ŋa ¹si-ma
#> 31                     ⁴camam ¹ŋa ¹si-ma, ¹ŋa ¹si-ma bel'-ri, ¹ŋa ¹si-ma
#> 32                     ⁴camam ¹ŋa ¹si-ma, ¹ŋa ¹si-ma bel'-ri, ¹ŋa ¹si-ma
#> 33                     ⁴camam ¹ŋa ¹si-ma, ¹ŋa ¹si-ma bel'-ri, ¹ŋa ¹si-ma
#> 34                     ⁴camam ¹ŋa ¹si-ma, ¹ŋa ¹si-ma bel'-ri, ¹ŋa ¹si-ma
#> 35                     ⁴camam ¹ŋa ¹si-ma, ¹ŋa ¹si-ma bel'-ri, ¹ŋa ¹si-ma
#> 36                     ⁴camam ¹ŋa ¹si-ma, ¹ŋa ¹si-ma bel'-ri, ¹ŋa ¹si-ma
#> 37                     ¹ŋa-la ¹sja, ²ai-ni-se, ³a-²khraŋ-na, ³a-⁴pjaŋ-na
#> 38                     ¹ŋa-la ¹sja, ²ai-ni-se, ³a-²khraŋ-na, ³a-⁴pjaŋ-na
#> 39                     ¹ŋa-la ¹sja, ²ai-ni-se, ³a-²khraŋ-na, ³a-⁴pjaŋ-na
#> 40                     ¹ŋa-la ¹sja, ²ai-ni-se, ³a-²khraŋ-na, ³a-⁴pjaŋ-na
#> 41                     ¹ŋa-la ¹sja, ²ai-ni-se, ³a-²khraŋ-na, ³a-⁴pjaŋ-na
#> 42                   ¹ŋa-la ¹sja, ²ai-ni ²kola-kat'-se ¹ca-o, ³pi-cim-ro
#> 43                   ¹ŋa-la ¹sja, ²ai-ni ²kola-kat'-se ¹ca-o, ³pi-cim-ro
#> 44                   ¹ŋa-la ¹sja, ²ai-ni ²kola-kat'-se ¹ca-o, ³pi-cim-ro
#> 45                   ¹ŋa-la ¹sja, ²ai-ni ²kola-kat'-se ¹ca-o, ³pi-cim-ro
#> 46                   ¹ŋa-la ¹sja, ²ai-ni ²kola-kat'-se ¹ca-o, ³pi-cim-ro
#> 47                   ¹ŋa-la ¹sja, ²ai-ni ²kola-kat'-se ¹ca-o, ³pi-cim-ro
#> 48                                       ³pi-ma, ³oca ¹ama ¹si-si ¹ni-ci
#> 49                                       ³pi-ma, ³oca ¹ama ¹si-si ¹ni-ci
#> 50                                       ³pi-ma, ³oca ¹ama ¹si-si ¹ni-ci
#> 51                                       ³pi-ma, ³oca ¹ama ¹si-si ¹ni-ci
#> 52                                       ³pi-ma, ³oca ¹ama ¹si-si ¹ni-ci
#> 53  ³ca ²som-se, ¹am-ta ¹phral-si, ²tha:-si, ¹am-ta ¹phral-si, ²tha:-si,
#> 54  ³ca ²som-se, ¹am-ta ¹phral-si, ²tha:-si, ¹am-ta ¹phral-si, ²tha:-si,
#> 55  ³ca ²som-se, ¹am-ta ¹phral-si, ²tha:-si, ¹am-ta ¹phral-si, ²tha:-si,
#> 56  ³ca ²som-se, ¹am-ta ¹phral-si, ²tha:-si, ¹am-ta ¹phral-si, ²tha:-si,
#> 57  ³ca ²som-se, ¹am-ta ¹phral-si, ²tha:-si, ¹am-ta ¹phral-si, ²tha:-si,
#> 58  ³ca ²som-se, ¹am-ta ¹phral-si, ²tha:-si, ¹am-ta ¹phral-si, ²tha:-si,
#> 59  ³ca ²som-se, ¹am-ta ¹phral-si, ²tha:-si, ¹am-ta ¹phral-si, ²tha:-si,
#> 60  ³ca ²som-se, ¹am-ta ¹phral-si, ²tha:-si, ¹am-ta ¹phral-si, ²tha:-si,
#> 61                                                      ³ca ¹thepa (aaa)
#> 62                                                      ³ca ¹thepa (aaa)
#> 63                                                      ³ca ¹thepa (aaa)
#> 64                                           ¹thepa ²paraŋ-ca ²tha:-ci, 
#> 65                                           ¹thepa ²paraŋ-ca ²tha:-ci, 
#> 66                                           ¹thepa ²paraŋ-ca ²tha:-ci, 
#> 67                          ³ca ¹cjaŋpa-ca ³kruŋ ²khru-pa-r' ²pit-cim-ro
#> 68                          ³ca ¹cjaŋpa-ca ³kruŋ ²khru-pa-r' ²pit-cim-ro
#> 69                          ³ca ¹cjaŋpa-ca ³kruŋ ²khru-pa-r' ²pit-cim-ro
#> 70                          ³ca ¹cjaŋpa-ca ³kruŋ ²khru-pa-r' ²pit-cim-ro
#> 71                          ³ca ¹cjaŋpa-ca ³kruŋ ²khru-pa-r' ²pit-cim-ro
#> 72                              ³ca ¹cjaŋpa-ca ³kruŋ ²khru-pa-ri ²pit-ma
#> 73                              ³ca ¹cjaŋpa-ca ³kruŋ ²khru-pa-ri ²pit-ma
#> 74                              ³ca ¹cjaŋpa-ca ³kruŋ ²khru-pa-ri ²pit-ma
#> 75                              ³ca ¹cjaŋpa-ca ³kruŋ ²khru-pa-ri ²pit-ma
#> 76                              ³ca ¹cjaŋpa-ca ³kruŋ ²khru-pa-ri ²pit-ma
#> 77                                                 ³kruŋ ²khru-si ³pa-ma
#> 78                                                 ³kruŋ ²khru-si ³pa-ma
#> 79                                                 ³kruŋ ²khru-si ³pa-ma
#> 80                               ³cjocjo ⁴ni:-se ¹tik ¹la-cim-ro ³pi-sam
#> 81                               ³cjocjo ⁴ni:-se ¹tik ¹la-cim-ro ³pi-sam
#> 82                               ³cjocjo ⁴ni:-se ¹tik ¹la-cim-ro ³pi-sam
#> 83                               ³cjocjo ⁴ni:-se ¹tik ¹la-cim-ro ³pi-sam
#> 84                               ³cjocjo ⁴ni:-se ¹tik ¹la-cim-ro ³pi-sam
#> 85                    ¹sja ²khraŋ-si, (bis) ³ol' ¹la-si, bag ²som ¹la-si
#> 86                    ¹sja ²khraŋ-si, (bis) ³ol' ¹la-si, bag ²som ¹la-si
#> 87                    ¹sja ²khraŋ-si, (bis) ³ol' ¹la-si, bag ²som ¹la-si
#> 88                    ¹sja ²khraŋ-si, (bis) ³ol' ¹la-si, bag ²som ¹la-si
#> 89                    ¹sja ²khraŋ-si, (bis) ³ol' ¹la-si, bag ²som ¹la-si
#> 90                    ¹sja ²khraŋ-si, (bis) ³ol' ¹la-si, bag ²som ¹la-si
#> 91                    ¹sja ²khraŋ-si, (bis) ³ol' ¹la-si, bag ²som ¹la-si
#> 92                    ¹sja ²khraŋ-si, (bis) ³ol' ¹la-si, bag ²som ¹la-si
#> 93                      ³cjocjo ⁴ni:-la ¹sja-cati ²tok ²sum ¹than-cim-ro
#> 94                      ³cjocjo ⁴ni:-la ¹sja-cati ²tok ²sum ¹than-cim-ro
#> 95                      ³cjocjo ⁴ni:-la ¹sja-cati ²tok ²sum ¹than-cim-ro
#> 96                      ³cjocjo ⁴ni:-la ¹sja-cati ²tok ²sum ¹than-cim-ro
#> 97                      ³cjocjo ⁴ni:-la ¹sja-cati ²tok ²sum ¹than-cim-ro
#> 98                      ³cjocjo ⁴ni:-la ¹sja-cati ²tok ²sum ¹than-cim-ro
#> 99                             ³kruŋ ²khru-si ¹kha-pa ¹cjaŋpa-ca-la ¹sja
#> 100                            ³kruŋ ²khru-si ¹kha-pa ¹cjaŋpa-ca-la ¹sja
#> 101                            ³kruŋ ²khru-si ¹kha-pa ¹cjaŋpa-ca-la ¹sja
#> 102                            ³kruŋ ²khru-si ¹kha-pa ¹cjaŋpa-ca-la ¹sja
#> 103                            ³kruŋ ²khru-si ¹kha-pa ¹cjaŋpa-ca-la ¹sja
#> 104                                « ¹ale, ¹ale, ¹in-che-m ³tolon ¹ca-ci
#> 105                                « ¹ale, ¹ale, ¹in-che-m ³tolon ¹ca-ci
#> 106                                « ¹ale, ¹ale, ¹in-che-m ³tolon ¹ca-ci
#> 107                                « ¹ale, ¹ale, ¹in-che-m ³tolon ¹ca-ci
#> 108                                « ¹ale, ¹ale, ¹in-che-m ³tolon ¹ca-ci
#> 109                                           ²ai-la bag ²cu-ri ¹mu-la »
#> 110                                           ²ai-la bag ²cu-ri ¹mu-la »
#> 111                                           ²ai-la bag ²cu-ri ¹mu-la »
#>                                                       sent_translation
#> 1                 autrefois, dans les temps anciens, là-haut, au Tibet
#> 2                 autrefois, dans les temps anciens, là-haut, au Tibet
#> 3                 autrefois, dans les temps anciens, là-haut, au Tibet
#> 4                 autrefois, dans les temps anciens, là-haut, au Tibet
#> 5                 autrefois, dans les temps anciens, là-haut, au Tibet
#> 6            autrefois, dans les temps anciens, là-haut, au Tibet donc
#> 7            autrefois, dans les temps anciens, là-haut, au Tibet donc
#> 8            autrefois, dans les temps anciens, là-haut, au Tibet donc
#> 9            autrefois, dans les temps anciens, là-haut, au Tibet donc
#> 10           autrefois, dans les temps anciens, là-haut, au Tibet donc
#> 11  il y avait un chetri, un brahmane et un tamang ; ils étaient trois
#> 12  il y avait un chetri, un brahmane et un tamang ; ils étaient trois
#> 13  il y avait un chetri, un brahmane et un tamang ; ils étaient trois
#> 14  il y avait un chetri, un brahmane et un tamang ; ils étaient trois
#> 15  il y avait un chetri, un brahmane et un tamang ; ils étaient trois
#> 16  il y avait un chetri, un brahmane et un tamang ; ils étaient trois
#> 17  il y avait un chetri, un brahmane et un tamang ; ils étaient trois
#> 18                                                     parmi ces trois
#> 19                           la mère était une vache, le père un singe
#> 20                           la mère était une vache, le père un singe
#> 21                           la mère était une vache, le père un singe
#> 22                           la mère était une vache, le père un singe
#> 23                                                   la mère leur dit:
#> 24                                                   la mère leur dit:
#> 25                                                   la mère leur dit:
#> 26                                          aux trois frères ses fils,
#> 27                                          aux trois frères ses fils,
#> 28                                          aux trois frères ses fils,
#> 29                       alors, quand je mourrai, à l'heure de ma mort
#> 30                       alors, quand je mourrai, à l'heure de ma mort
#> 31                       alors, quand je mourrai, à l'heure de ma mort
#> 32                       alors, quand je mourrai, à l'heure de ma mort
#> 33                       alors, quand je mourrai, à l'heure de ma mort
#> 34                       alors, quand je mourrai, à l'heure de ma mort
#> 35                       alors, quand je mourrai, à l'heure de ma mort
#> 36                       alors, quand je mourrai, à l'heure de ma mort
#> 37          ma chair, vous autres, ne l'incinérer pas, ne la jetez pas
#> 38          ma chair, vous autres, ne l'incinérer pas, ne la jetez pas
#> 39          ma chair, vous autres, ne l'incinérer pas, ne la jetez pas
#> 40          ma chair, vous autres, ne l'incinérer pas, ne la jetez pas
#> 41          ma chair, vous autres, ne l'incinérer pas, ne la jetez pas
#> 42                  ma chair, vous autres enfants, mangez-la, dit-elle
#> 43                  ma chair, vous autres enfants, mangez-la, dit-elle
#> 44                  ma chair, vous autres enfants, mangez-la, dit-elle
#> 45                  ma chair, vous autres enfants, mangez-la, dit-elle
#> 46                  ma chair, vous autres enfants, mangez-la, dit-elle
#> 47                  ma chair, vous autres enfants, mangez-la, dit-elle
#> 48                                    avec ces paroles, la mère mourut
#> 49                                    avec ces paroles, la mère mourut
#> 50                                    avec ces paroles, la mère mourut
#> 51                                    avec ces paroles, la mère mourut
#> 52                                    avec ces paroles, la mère mourut
#> 53           les trois fils dépecèrent leur mère, la découpèrent (bis)
#> 54           les trois fils dépecèrent leur mère, la découpèrent (bis)
#> 55           les trois fils dépecèrent leur mère, la découpèrent (bis)
#> 56           les trois fils dépecèrent leur mère, la découpèrent (bis)
#> 57           les trois fils dépecèrent leur mère, la découpèrent (bis)
#> 58           les trois fils dépecèrent leur mère, la découpèrent (bis)
#> 59           les trois fils dépecèrent leur mère, la découpèrent (bis)
#> 60           les trois fils dépecèrent leur mère, la découpèrent (bis)
#> 61                                                  le fils aîné (euh)
#> 62                                                  le fils aîné (euh)
#> 63                                                  le fils aîné (euh)
#> 64                                   l'aîné et le second découpèrent, 
#> 65                                   l'aîné et le second découpèrent, 
#> 66                                   l'aîné et le second découpèrent, 
#> 67                    et ils envoyèrent le plus jeune laver les boyaux
#> 68                    et ils envoyèrent le plus jeune laver les boyaux
#> 69                    et ils envoyèrent le plus jeune laver les boyaux
#> 70                    et ils envoyèrent le plus jeune laver les boyaux
#> 71                    et ils envoyèrent le plus jeune laver les boyaux
#> 72                       ils envoyèrent le plus jeune laver les boyaux
#> 73                       ils envoyèrent le plus jeune laver les boyaux
#> 74                       ils envoyèrent le plus jeune laver les boyaux
#> 75                       ils envoyèrent le plus jeune laver les boyaux
#> 76                       ils envoyèrent le plus jeune laver les boyaux
#> 77                               et quand il rapporta les boyaux lavés
#> 78                               et quand il rapporta les boyaux lavés
#> 79                               et quand il rapporta les boyaux lavés
#> 80                                   que firent ses deux grands frères
#> 81                                   que firent ses deux grands frères
#> 82                                   que firent ses deux grands frères
#> 83                                   que firent ses deux grands frères
#> 84                                   que firent ses deux grands frères
#> 85     ayant rôti la viande, (bis) comme ça, ils en firent trois parts
#> 86     ayant rôti la viande, (bis) comme ça, ils en firent trois parts
#> 87     ayant rôti la viande, (bis) comme ça, ils en firent trois parts
#> 88     ayant rôti la viande, (bis) comme ça, ils en firent trois parts
#> 89     ayant rôti la viande, (bis) comme ça, ils en firent trois parts
#> 90     ayant rôti la viande, (bis) comme ça, ils en firent trois parts
#> 91     ayant rôti la viande, (bis) comme ça, ils en firent trois parts
#> 92     ayant rôti la viande, (bis) comme ça, ils en firent trois parts
#> 93                    les deux grands frères cachèrent vite leur part 
#> 94                    les deux grands frères cachèrent vite leur part 
#> 95                    les deux grands frères cachèrent vite leur part 
#> 96                    les deux grands frères cachèrent vite leur part 
#> 97                    les deux grands frères cachèrent vite leur part 
#> 98                    les deux grands frères cachèrent vite leur part 
#> 99        et la part du jeune frère qui revenait ayant lavé les boyaux
#> 100       et la part du jeune frère qui revenait ayant lavé les boyaux
#> 101       et la part du jeune frère qui revenait ayant lavé les boyaux
#> 102       et la part du jeune frère qui revenait ayant lavé les boyaux
#> 103       et la part du jeune frère qui revenait ayant lavé les boyaux
#> 104                      « frère, frère, nous autres avons mangé avant
#> 105                      « frère, frère, nous autres avons mangé avant
#> 106                      « frère, frère, nous autres avons mangé avant
#> 107                      « frère, frère, nous autres avons mangé avant
#> 108                      « frère, frère, nous autres avons mangé avant
#> 109                                                 ta part est ici »,
#> 110                                                 ta part est ici »,
#> 111                                                 ta part est ici »,
#>  [ reached 'max' / getOption("max.print") -- omitted 229 rows ]