Rename rezrDF columns.

Rename rezrDF columns.

Usage

rez_rename(df, ...)

Arguments

df

The data frame to be modified.

...

Functions to be passed to rename. New column names are argument names; old column names are argument values.

Value

A rezrDF object.

Note

This function does not update foreign references to the field that you're renaming. So be sure to update the updateFunctions of those fields; otherwise, you will break your rezrObj.

Examples

sbc007$treeEntryDF$default %>% rez_rename(gramRel = Relation)
#> Tip: When performed on a rezrDF inside a rezrObj, rez_rename is a potentially destructive action. It is NOT recommended to assign it back to a rezrDF inside a rezrObj. If you must do so, be careful to update all addresses from other DFs to this DF.
#> # A tibble: 1,525 × 16
#>    id        doc   tree  order sourceLink level tokenOrderFirst docTokenSeqFirst
#>    <chr>     <chr> <chr> <dbl> <chr>      <dbl>           <dbl>            <dbl>
#>  1 3419C1EF… sbc0… 1235…     0 ""            -1               1                1
#>  2 272B7E44… sbc0… 1235…     1 ""            -1               2                2
#>  3 9403F199… sbc0… 1235…     2 ""            -1               3                3
#>  4 1FA81754… sbc0… 36F1…     0 ""            -1               1                4
#>  5 E5564F08… sbc0… 1235…     3 "D98B08C2…     1               1                4
#>  6 87560FEB… sbc0… 79D2…     0 "D4796592…     1               1                4
#>  7 2198692B… sbc0… 36F1…     1 ""            -1               2                5
#>  8 29A7D8B3… sbc0… 79D2…     1 ""             0               2                5
#>  9 3005263A… sbc0… 1235…     4 ""             0               2                5
#> 10 163E2115… sbc0… 36F1…     2 "2234E289…     1               3                6
#> # ℹ 1,515 more rows
#> # ℹ 8 more variables: tokenOrderLast <dbl>, docTokenSeqLast <dbl>, text <chr>,
#> #   transcript <chr>, parent <chr>, subtype <chr>, gramRel <chr>, layer <chr>