Read and write rezrDFs as CSV files

Replacement of readr read_csv and write_csv, but contains added functionality for dealing with rezrDFs more easily.

rez_write_csv(df, path, inclCols = character(0), exclCols = character(0), ...)

rez_read_csv(
  path,
  origDF = NULL,
  lubridate = F,
  inclCols = character(0),
  exclCols = character(0),
  ...
)

Arguments

df

The rezrDF to be exported.

path

The path from which a CSV is to be imported / exported.

inclCols

Columns to be included from the import / export.

exclCols

Columns to be excluded from the import / export.

...

Arguments passed onto read_csv and write_csv.

origDF

The rezrDF that originally produced this CSV, used to identify data types of the columns.

lubridate

Are you using the lubridate package?

Value

For rez_read_csv, a data.frame for the CSV being imported. This is NOT a rezrDF, so please do not assign it to a rezrObj; use updateFromDF to update an existing rezrDF with the imported data.frame.